pyproject.toml
[tool.poetry]
name = "research-analysis-suite"
version = "0.1.0"
description = "A modern research toolkit powered by AI, LLMs, and statistical analysis."
authors = ["Cristian Restrepo"]
readme = "README.md"
packages = [
    { include = "research_methods" },
    { include = "utils" }
]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
openai = "^1.73.0"
python-dotenv = "^1.1.0"
pydub = "^0.25.1"
imageio-ffmpeg = "^0.4.8"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
